refactor: drop namespace proof-of-concept code changes
