+++ packages/opencode/src/cli/cmd/debug/config.ts
@@ -1,5 +1,5 @@
 import { EOL } from "os"
-import { Config } from "@/config/config"
+import { Config } from "../../../config"
 import { AppRuntime } from "@/effect/app-runtime"
 import { bootstrap } from "../../bootstrap"
 import { cmd } from "../cmd"
