+++ packages/opencode/src/cli/cmd/tui/worker.ts
@@ -5,7 +5,7 @@ import { Instance } from "@/project/instance"
 import { InstanceBootstrap } from "@/project/bootstrap"
 import { Rpc } from "@/util"
 import { upgrade } from "@/cli/upgrade"
-import { Config } from "@/config/config"
+import { Config } from "@/config"
 import { GlobalBus } from "@/bus/global"
 import { Flag } from "@/flag/flag"
 import { writeHeapSnapshot } from "node:v8"
