+++ packages/opencode/src/file/watcher.ts
@@ -12,7 +12,7 @@ import { Flag } from "@/flag/flag"
 import { Git } from "@/git"
 import { Instance } from "@/project/instance"
 import { lazy } from "@/util/lazy"
-import { Config } from "@/config/config"
+import { Config } from "../config"
 import { FileIgnore } from "./ignore"
 import { Protected } from "./protected"
 import { Log } from "../util"
