+++ packages/opencode/src/server/instance/config.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
-import { Config } from "@/config/config"
+import { Config } from "../../config"
 import { Provider } from "../../provider"
 import { mapValues } from "remeda"
 import { errors } from "../error"
