+++ packages/opencode/src/server/instance/provider.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
-import { Config } from "@/config/config"
+import { Config } from "../../config"
 import { Provider } from "../../provider"
 import { ModelsDev } from "../../provider"
 import { ProviderAuth } from "../../provider"
