+++ packages/opencode/src/tool/task.ts
@@ -6,7 +6,7 @@ import { SessionID, MessageID } from "../session/schema"
 import { MessageV2 } from "../session/message-v2"
 import { Agent } from "../agent/agent"
 import type { SessionPrompt } from "../session/prompt"
-import { Config } from "@/config/config"
+import { Config } from "../config"
 import { Effect } from "effect"
 
 export interface TaskPromptOps {
