+++ packages/opencode/test/config/agent-color.test.ts
@@ -3,7 +3,7 @@ import { Effect } from "effect"
 import path from "path"
 import { provideInstance, tmpdir } from "../fixture/fixture"
 import { Instance } from "../../src/project/instance"
-import { Config } from "../../src/config/config"
+import { Config } from "../../src/config"
 import { Agent as AgentSvc } from "../../src/agent/agent"
 import { Color } from "../../src/util"
 import { AppRuntime } from "../../src/effect/app-runtime"
