+++ packages/opencode/test/permission-task.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, test, expect } from "bun:test"
 import { Permission } from "../src/permission"
-import { Config } from "../src/config/config"
+import { Config } from "../src/config"
 import { Instance } from "../src/project/instance"
 import { tmpdir } from "./fixture/fixture"
 import { AppRuntime } from "../src/effect/app-runtime"
